adk-genie-mcp-agent/pyproject.toml
[project]
name = "adk-genie-mcp-agent"
version = "0.1.0"
description = "Google ADK agent connecting to Databricks AI/BI Genie via MCP, deployed to Vertex AI Agent Engine"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "databricks-sdk>=0.40.0",
    "openai>=1.30.0",
]

[project.optional-dependencies]
deploy = [
    "google-cloud-aiplatform[agent_engines,adk]>=1.88.0",
    "cloudpickle>=3.0.0",
]
dev = [
    "google-cloud-aiplatform[agent_engines,adk]>=1.88.0",
    "cloudpickle>=3.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
